Move out PodmanRunner obj, fix UT
